+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### [2.0.1-dev.1] - Feb 15, 2026
 
-- [BREAKING] Fix: [Issue: #256](https://github.com/maheshj01/searchfield/issues/256) Add maintainLabelSize to SearchInputDecoration (Requires Flutter 3.41.0 or greater)
+- [BREAKING] Fix: [Issue: #256](https://github.com/maheshj01/searchfield/issues/256) Add `maintainLabelSize` and `enabled` to SearchInputDecoration (Requires Flutter 3.41.0 or greater)
 
 ### [1.4.0-dev.5] - Nov 01, 2025